+++ src/app/inventory/dimCsvService.factory.ts
@@ -25,6 +25,8 @@ const FILTER_NODE_NAMES = [
   'No Projection'
 ];
 
+const events = ['', 'Dawning', 'Crimson Days', 'Solstice of Heroes', 'Festival of the Lost'];
+
 function capitalizeFirstLetter(str: string) {
   if (!str || str.length === 0) {
     return '';
@@ -105,8 +107,8 @@ function downloadArmor(items, nameMap) {
         'Locked,Equipped,Year,DTR Rating,# of Reviews,% Quality,' +
         '% IntQ,% DiscQ,% StrQ,Int,Disc,Str,Notes,Perks\n'
       : 'Name,Hash,Id,Tag,Tier,Type,Equippable,Power,Owner,Locked,' +
-        'Equipped,Year,DTR Rating,# of Reviews,Mobility,Recovery,' +
-        'Resilience,Notes,Perks\n';
+        'Equipped,Year,Season,Event,DTR Rating,# of Reviews,Mobility,' +
+        'Recovery,Resilience,Notes,Perks\n';
   let data = '';
   items.forEach((item) => {
     data += `"${item.name}",`;
@@ -126,7 +128,10 @@ function downloadArmor(items, nameMap) {
     data += item.destinyVersion === 1 ? `${(item.percentComplete * 100).toFixed(0)},` : ``;
     data += `${item.locked},`;
     data += `${item.equipped},`;
-    data += item.destinyVersion === 1 ? `${item.year},` : item.season < 3 ? `1,` : `2,`;
+    data += item.destinyVersion === 1 ? `${item.year},` : item.season <= 3 ? `1,` : `2,`;
+    data += item.destinyVersion === 1 ? '' : `${item.season},`;
+    data +=
+      item.destinyVersion === 1 ? '' : item.event ? `${events[item.event]},` : `${events[0]},`;
     if (item.dtrRating && item.dtrRating.overallScore) {
       data += `${item.dtrRating.overallScore},${item.dtrRating.ratingCount},`;
     } else {
@@ -195,8 +200,8 @@ function downloadWeapons(guns, nameMap) {
         'AA,Impact,Range,Stability,ROF,Reload,Mag,Equip,' +
         'Notes,Nodes\n'
       : 'Name,Hash,Id,Tag,Tier,Type,Power,Dmg,Owner,' +
-        'Locked,Equipped,Year,DTR Rating,# of Reviews,' +
-        'AA,Impact,Range,Stability,ROF,Reload,Mag,Equip,' +
+        'Locked,Equipped,Year,Season,Event,DTR Rating,' +
+        '# of Reviews,AA,Impact,Range,Stability,ROF,Reload,Mag,Equip,' +
         'Notes,Nodes\n';
   let data = '';
   guns.forEach((gun) => {
@@ -217,6 +222,8 @@ function downloadWeapons(guns, nameMap) {
     data += `${gun.locked},`;
     data += `${gun.equipped},`;
     data += gun.destinyVersion === 1 ? `${gun.year},` : gun.season <= 3 ? `1,` : `2,`;
+    data += gun.destinyVersion === 1 ? '' : `${gun.season},`;
+    data += gun.destinyVersion === 1 ? '' : gun.event ? `${events[gun.event]},` : `${events[0]},`;
     if (gun.dtrRating && gun.dtrRating.overallScore) {
       data += `${gun.dtrRating.overallScore},${gun.dtrRating.ratingCount},`;
     } else {